Update status sheet script
The status sheet script depends on some parts of the main analysis script. Since it was refactored, this updates the status sheet script to call all the methods correctly.
